refactor(lsp): split helpers into lsp/helpers.rs to restore branch coverage
Move all sync functions (build_rule_index, detect_completion_context,
severity_completions, extract_rule_id_at, find_workspace_dir,
to_lsp_diagnostic, lint_file_diagnostics) plus their 30 unit tests
into a new lsp/helpers.rs submodule.

lsp.rs now contains only the async Backend, LanguageServer impl, and
run() entry point. The --ignore-filename-regex 'lsp\.rs' in CI still
excludes lsp.rs (avoiding the LLVM 22.1 getInstantiationGroups SIGSEGV
on async state machines), but lsp/helpers.rs is now counted, restoring
branch coverage from 88.53% to 93.02%.
